LPD-34745 Moves logic since those methods are only used in siteTemplatesWithWebContent.spec.ts
